Kotlin rewrite
The main motivation for moving to Kotlin was initially the
nonnull/nullable being directly part of the language. It's really not
practical to use Eclipse's annotations inside a non-Eclipse project.

Kotlin seems to offer a ton of other features, like sealed class that
will be a perfect fit for field value types.

Also bring in the notion of trace collections, backed by a generic
SortedCompoundIterator.

Signed-off-by: Alexandre Montplaisir <alexmonthy@efficios.com>