ctf: Do not use <goal>test-jar</goal> after all
While it was flaky but working in a deployed Maven repo, with a p2 repo
it compeltely breaks down: the p2 repo ONLY contains the test-jar, not
the real jar! (bug in Eclipse Package Drone?).

Revert the usage of the test-jar <goal> in the ctf module, and just
duplicate the needed file in the lttng one.

Bump version to 0.2.7-SNAPSHOT.

Signed-off-by: Alexandre Montplaisir <alexmonthy@efficios.com>