Settle on GSON for the JSON output example
org.json has crazy version numbers (the full date being actually the
MAJOR number), so when converted to an OSGi package it prevents any
other version from being used.

Since of course that exact version is not present in Orbit, it's
impossible for Eclipse clients to find a match.

Settle on GSON for the simple reason that is has a saner versioning
policy.

Version 0.2.10.

Signed-off-by: Alexandre Montplaisir <alexmonthy@efficios.com>