core: Allow tracking state within a StateSystemAnalysis' execution
It is not uncommon for state system analyses to track temporary state
during the construction of a state system. This state used to be part of
the analysis object directly.

Since analyses are singletons now, these singletons should NOT hold such
state! Instead, add a overridable method which will specify an array of
objects that will be tracked within a single call to execute() (and sent
to every eventHandle() call). This is not the prettiest solution, but it
should allow analyses such as the UST debug info to actually work.

Bump to version 0.2.17.

Signed-off-by: Alexandre Montplaisir <alexmonthy@efficios.com>