timegraph: Reduce total number of state system queries
Instead of letting createInterval() do its own ss queries, first ask the
model implementation to provide which quarks it would like to know
about, then pass it back the results in a separate abstract method.

This will allow the base class to aggregate all the requested quarks,
and query them all in one go with one call to queryStates(), potentially
avoiding a lot of smaller queries all the same timestamp.

With this and the previous commit, ThreadsModelBenchmark on my HDD
machine with a large reference trace goes from 44s to ~5s! :O

Signed-off-by: Alexandre Montplaisir <alexmonthy@efficios.com>