Relax check in EventStatsXYChartProvider
If the requested time range begins before the actual state system,
simply clamp the request to the valid range instead of throwing an
exception.

Clients mostly don't have direct access to the state system (and its
bounds) anyway, but rather just the trace project. In case of
multi-trace projects, it's important we don't fail if trace-specific
state systems are used.

Signed-off-by: Alexandre Montplaisir <alexmonthy@efficios.com>