core: Add TraceIterator seek API
Since trace iterators are analoguous to (and usually implemented using)
file descriptors, it would be convenient to be able to seek() an
existing iterator, so that we don't have to close() the old one and
re-open a new one.

This means that implementations will have to wrap Java Iterators so that
they can carry that extra state.

Signed-off-by: Alexandre Montplaisir <alexmonthy@efficios.com>