core: Introduce the notion of XYChart series providers
A XY Chart *series* provider will handle a single series of an XYChart.
It will regroup the series (immutable) object itself, and the logic of
how to generate renders of this series. It will also track the usual
enabled/disabled flag, which was present in timegraph series but not in
xy-chart.

The XY Chart *model* provider will aggregate different series providers.
It can supply its own, and allow external component to register new
ones. This makes it more similar to the TimeGraphModel provider, with
its different "layers" for different data providers.

Signed-off-by: Alexandre Montplaisir <alexmonthy@efficios.com>